feat: implement extension track cache, source selection, and add comprehensive unit tests

+++ core/shared/src/main/java/com/theveloper/pixelplay/shared/WearLyrics.kt
@@ -13,7 +13,7 @@ data class WearLyrics(
 
 @Serializable
 data class WearSyncedLyricLine(
-    val timeMs: Int,
+    val timeMs: Long,
     val line: String,
     val translation: String? = null,
     val romanization: String? = null,
